Add files via upload
The provided Jupyter Notebook (ipynb file) conducts a comprehensive analysis of COVID-19 data using Python. The dataset is sourced from various files, including Worldometer data and country-wise statistics, and is organized into several dataframes for detailed examination. The analysis covers a wide range of aspects, starting with the preparation of data, cleaning, and exploratory data analysis (EDA). Descriptive statistics are generated for key features, and spatial analysis is performed to identify the countries of origin for hotel guests.

The notebook then delves into the analysis of room reservations and assignments, market segment bookings, and patterns in guest arrivals over time. Visualizations, such as choropleth maps and bar plots, are used to effectively communicate insights. The code also includes explanations and comments to enhance understanding.

The latter part of the notebook focuses on COVID-19 data analysis. It explores total cases, deaths, recoveries, and active cases globally and in the top 20 affected countries. Visualizations like treemaps, line plots, and bar plots aid in understanding trends over time and country-specific impacts. In-depth analyses are performed on death rates, recovery rates, and test ratios.

The notebook concludes with a series of visualizations that provide a holistic view of the COVID-19 situation, including donut charts, death-confirmed ratios, recovery-confirmed ratios, and serious to death case rates. Additionally, a user-defined function facilitates country-wise visualizations for a more detailed examination of specific regions. The README file accompanying the notebook provides comprehensive guidance for replicating the analysis and encourages contributions and improvements to the project.